requirements.txt
functions-framework==3.*

numpy~=1.24.3
torch~=2.0.0
opencv-python~=4.7.0.72
scipy~=1.10.1
tqdm~=4.65.0
torchvision~=0.15.1
fastai~=1.0.60
matplotlib~=3.7.1
spacy~=3.5.2
fastprogress~=1.0.3
Pillow~=9.5.0
nbconvert~=7.3.1
ipython~=8.12.0
nbformat~=5.8.0
traitlets~=5.9.0
PyYAML~=6.0
pandas~=2.0.1
requests~=2.28.2
beautifulsoup4~=4.12.2
ffmpeg~=1.4
setuptools~=58.1.0
Flask~=2.3.1